requirements.txt
anthropic>=0.39.0
pydantic>=2.0.0
loguru>=0.7.0
apscheduler>=3.10.0
python-telegram-bot>=21.0
httpx>=0.27.0
pandas>=2.0.0
pandas-ta>=0.3.14b0
python-dotenv>=1.0.0
python-binance>=1.0.19
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
